% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clusterProjDivisive}
\alias{clusterProjDivisive}
\title{Divisive (hierarchical) clustering on the projective space}
\usage{
clusterProjDivisive(X, tol, iter.max = 100)
}
\arguments{
\item{X}{the data belonging to the projective space}

\item{tol}{the tolerance that when reached, stops increasing the number of clusters.
At each step, the (change in wcss) / (original wcss) must be above this tolerance.
In general, as the tolerance decreases, the number of clusters in the output increases.}

\item{iter.max}{the maximum number of iterations}
}
\value{
A list with the following components:
\itemize{
         \item{c} {Vector of real numbers from 1 to K representing the cluster
                     that the corresponding X value belongs to.}
         \item{sseIndividual} {Vector of the within sum-of-squares for each cluster}
         \item{rmse} {The total within-cluster root mean-squared-error for the
         obtained cluster} 
         \item{rmseSequence} {The change in total within-cluster root mean-squared-error for each
                             iteration of the function}
}
}
\description{
Creates clusters of points on the projective space using divisive k-means clustering
}
\examples{
n1 <- 37; n2 <- 19
x1 <- rnorm(n1, 6); y1 <- rnorm(n1, 0); z1 <- rnorm(n1, 0, 0.1)
x2 <- rnorm(n2, 8); y2 <- rnorm(n2, 8); z2 <- rnorm(n2, 0, 0.1)
X <- rbind(cbind(x1, y1, z1), cbind(x2, y2, z2)) * sample(c(-1, 1), size=n1+n2, replace=TRUE)
X <- X / sqrt(rowSums(X^2))
(c <- clusterProjDivisive(X=X, tol=0.1))

}
\seealso{
clusterProjKmeans
}
\author{
Paul Smith, \email{mmpws@leeds.ac.uk}
}
\keyword{clustering,}
\keyword{kmeans}
