% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{entropy}
\alias{entropy}
\title{Entropy approximation using the m-spacing method}
\usage{
entropy(x, m)
}
\arguments{
\item{x}{the data, either a vector or matrix with columns 
representing dimension}

\item{m}{(optional) the m-spacing. Defaults to m <- sqrt(nrow(x)) 
if missing}
}
\value{
Vector of real numbers corresponding to the approximate 
             entropy for each row of input x.
}
\description{
Calculates entropy using m-spacing, as in 
Beirlant, Jan, et al. "Nonparametric entropy estimation: An overview."
}
\examples{
X1 <- matrix(rnorm(150), ncol=15, nrow=10)
X2 <- matrix(rnorm(150), ncol=10, nrow=15)
X3 <- matrix(rnorm(1500), ncol=10, nrow=150)
X4 <- matrix(rnorm(1500), ncol=100, nrow=15)
X <- list(X1, X2, X3, X4)
Xi.entr <- vector("list", length = length(X))
for (i in 1:length(X)) {
    Xi <- X[[i]] 
    Xi.entr_mat <- entropy(Xi)
    Xi.entr[[i]] <- Xi.entr_mat
}
str(Xi.entr)

}
\references{
\url{https://www.researchgate.net/profile/Jan_Beirlant/publication/243768404_Nonparametric_Entropy_Estimation_An_Overview/links/0046352937863d9023000000/Nonparametric-Entropy-Estimation-An-Overview.pdf}
}
\author{
Paul Smith \& Jochen Voss, \email{mmpws@leeds.ac.uk}
}
\keyword{entropy}
