% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{projective.cluster}
\alias{projective.cluster}
\title{Kmeans clustering on the projective space}
\usage{
projective.cluster(X, K, maxiter = 100, initial, verbose = TRUE)
}
\arguments{
\item{X}{the data belonging to the projective space}

\item{K}{the number of clusters required in output}

\item{maxiter}{the maximum number of iterations}

\item{initial}{(optional) the initial clustering. 
If missing(K), then initial clusters found using the k-means++ method}

\item{verbose}{gives information of completed clusters}
}
\value{
Vector of real numbers from 1 to K representing the cluster
    that the corresponding X value belongs to.
}
\description{
Creates K clusters of points on the projective space using the k-means method
}
\examples{
n1 <- 37; n2 <- 19
x1 <- rnorm(n1, 6); y1 <- rnorm(n1, 0); z1 <- rnorm(n1, 0, 0.1)
x2 <- rnorm(n2, 8); y2 <- rnorm(n2, 8); z2 <- rnorm(n2, 0, 0.1)
X <- rbind(cbind(x1, y1, z1), cbind(x2, y2, z2)) * sample(c(-1, 1), size=n1+n2, replace=TRUE)
X <- X / sqrt(rowSums(X^2))
(c <- projective.cluster(X, 2))

}
\seealso{
\code{\link{projective.divisive_clust}}
}
\author{
Jochen Voss, \email{Jochen.Voss@leeds.ac.uk}
}
\keyword{clustering,}
\keyword{kmeans}
