% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{projective.divisive_clust}
\alias{projective.divisive_clust}
\title{Divisive (hierarchical) clustering on the projective space}
\usage{
projective.divisive_clust(X, tol, maxiter = 100)
}
\arguments{
\item{X}{the data belonging to the projective space}

\item{tol}{the tolerance that when reached, stops increasing the number of clusters. 
At each step, the (change in wss) / (original wss) must be above this tolerance}

\item{maxiter}{the maximum number of iterations}
}
\value{
A list with the following components:
\itemize{
         \item{c} {Vector of real numbers from 1 to K representing the cluster
                     that the corresponding X value belongs to.}
         \item{rss}{Vector of the within sum-of-squares for each cluster}
         \item{wss}{The total within sum-of-squares for the outputted cluster}
         \item{wss_all}{The change in total within sum-of-squares for each 
                             iteration of the function}
}
}
\description{
Creates clusters of points on the projective space using divisive kmeans clustering
}
\examples{
n1 <- 37; n2 <- 19
x1 <- rnorm(n1, 6); y1 <- rnorm(n1, 0); z1 <- rnorm(n1, 0, 0.1)
x2 <- rnorm(n2, 8); y2 <- rnorm(n2, 8); z2 <- rnorm(n2, 0, 0.1)
X <- rbind(cbind(x1, y1, z1), cbind(x2, y2, z2)) * sample(c(-1, 1), size=n1+n2, replace=TRUE)
X <- X / sqrt(rowSums(X^2))
(c <- projective.divisive_clust(X=X, tol=0.1))

}
\seealso{
projective.cluster
}
\author{
Paul Smith, \email{mmpws@leeds.ac.uk}
}
\keyword{clustering,}
\keyword{kmeans}
