% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Entropy estimation using m-spacing}
\usage{
entropy(x, m)
}
\arguments{
\item{x}{the data, either a vector or matrix.
If x is a matrix, entropy is estimated for each row separately.}

\item{m}{(optional) the m-spacing. Defaults to m <- sqrt(n)
if missing, where n is length(x) if x is a vector,
    or ncol(x) if a matrix}
}
\value{
Vector of real numbers corresponding to the approximate
             entropy for each row of input x.
}
\description{
Calculates entropy using m-spacing.
}
\examples{
X1 <- matrix(rnorm(150), ncol=15, nrow=10)
X2 <- matrix(rnorm(150), ncol=10, nrow=15)
X3 <- matrix(rnorm(1500), ncol=10, nrow=150)
X4 <- matrix(rnorm(1500), ncol=100, nrow=15)
X <- list(X1, X2, X3, X4)
XiEntr <- vector("list", length = length(X))
for (i in 1:length(X)) {
    Xi <- X[[i]]
    XiEntr_mat <- entropy(Xi)
    XiEntr[[i]] <- XiEntr_mat
}
str(XiEntr)

}
\references{
Beirlant, Jan, et al. "Nonparametric entropy estimation: An overview."
}
\author{
Paul Smith \& Jochen Voss, \email{mmpws@leeds.ac.uk}
}
\keyword{entropy}
